Verify credentials on requests that access or modify user data (#84)
* initial commit

* added authentication chacks for each of the endpoints that need it

---------

+++ src/main/java/main/controller/CartController.java
@@ -2,6 +2,7 @@ package main.controller;
 
 import main.model.UserModel;
 import main.repository.UserRepository;
+import main.model.Credentials;
 import main.service.UserService;
 import org.springframework.beans.factory.annotation.Autowired;
 import org.springframework.http.MediaType;
@@ -27,17 +28,22 @@ public class CartController {
         this.userRepository = userRepository;
     }
 
-
-    @GetMapping(value="/getUserCart", produces= MediaType.APPLICATION_JSON_VALUE)
-    public ArrayList<String[]> getBooks(@RequestParam("userid") Long userid) {
-        UserModel user = userRepository.findAllById(Arrays.asList(userid)).get(0);
-        Map<BookModel, Integer> Books = user.getShoppingCart().getBookQuantityMap();
+    @CrossOrigin
+    @PostMapping(value="/getUserCart", produces= MediaType.APPLICATION_JSON_VALUE)
+    public ArrayList<String[]> getBooks(@RequestParam("userid") Long userid, @RequestBody Credentials credentials) {
         ArrayList<String[]> bookList = new ArrayList<>();
-        for (Map.Entry<BookModel, Integer> entry : Books.entrySet()) {
-            String[] arr = {entry.getKey().getTitle(), entry.getKey().getAuthor(), entry.getKey().getPrice().toString(), entry.getValue().toString(),  entry.getKey().getId().toString()};
-            bookList.add(arr);
+        boolean valid = userService.verifyLogin(credentials.getUsername(), credentials.getPassword());
+        
+        if(valid){
+            UserModel user = userRepository.findAllById(Arrays.asList(userid)).get(0);
+            Map<BookModel, Integer> Books = user.getShoppingCart().getBookQuantityMap();
+            for (Map.Entry<BookModel, Integer> entry : Books.entrySet()) {
+                String[] arr = {entry.getKey().getTitle(), entry.getKey().getAuthor(), entry.getKey().getPrice().toString(), entry.getValue().toString(),  entry.getKey().getId().toString()};
+                bookList.add(arr);
+            }
         }
-        return bookList;
+
+        return bookList; // empty could mean invalid credentials
     }
 
     /**
@@ -46,32 +52,49 @@ public class CartController {
      * @param bookId
      * @return
      */
+    @CrossOrigin
     @PostMapping("/addToCart")
-    public String addBook(@RequestParam("userid") Long userid, @RequestParam("bookId") Long bookId, @RequestParam("quantity") int quantity) {
-        boolean addTo = userService.addToCart(userid, bookId, quantity);
-        if (addTo){
-            return  "{\"success\": true}";
+    public String addBook(@RequestParam("userid") Long userid, @RequestParam("bookId") Long bookId, @RequestParam("quantity") int quantity, @RequestBody Credentials credentials) {
+        boolean valid = userService.verifyLogin(credentials.getUsername(), credentials.getPassword());
+
+        if(valid){    
+            boolean addTo = userService.addToCart(userid, bookId, quantity);
+            if (addTo){
+                return  "{\"success\": true}";
+            }
         }
-        return  "{\"success\": false}";
+
+        return  "{\"success\": false}"; // could be because of invalid credentials
     }
 
+    @CrossOrigin
     @DeleteMapping("/deleteFromCart")
-    public String deleteCartItem(@RequestParam("userid") Long userId, @RequestParam("bookId") Long bookId) {
-        System.out.print("testing cartController");
-        boolean removeFrom = userService.removeFromCart(userId, bookId);
-        if (removeFrom){
-            return "{\"success\": true}";
+    public String deleteCartItem(@RequestParam("userid") Long userId, @RequestParam("bookId") Long bookId, @RequestBody Credentials credentials) {
+        boolean valid = userService.verifyLogin(credentials.getUsername(), credentials.getPassword());
+
+        if(valid){
+            boolean removeFrom = userService.removeFromCart(userId, bookId);
+            if (removeFrom){
+                return "{\"success\": true}";
+            }
         }
-        return "{\"success\": false}";
+
+        return "{\"success\": false}"; // could be because of invalid credentials
     }
 
+    @CrossOrigin
     @PostMapping("/checkoutCart")
-    public String checkout(@RequestParam("userid") Long userid) {
-        boolean orderMade = userService.createOrder(userid);
-        if (orderMade){
-            return  "{\"success\": true}";
+    public String checkout(@RequestParam("userid") Long userid, @RequestBody Credentials credentials) {
+        boolean valid = userService.verifyLogin(credentials.getUsername(), credentials.getPassword());
+
+        if(valid){
+            boolean orderMade = userService.createOrder(userid);
+            if (orderMade){
+                return  "{\"success\": true}";
+            }
         }
-        return  "{\"success\": false}";
+
+        return  "{\"success\": false}"; // could be because of invalid credentials
     }
 
 }
